230214/정선아/#4까지 수정
1. EveningScene 영수증
2. EveningScene 3일마다 월세 차감
3. EveningScene 저녁구매팝업
   - 재료팝업 구현 완료(확률계산, 판매완료, 목록갱신)
4. 1-3번 끝나면 NightScene으로 이동

+++ Assets/Script/Class/GameData/Ingredient.cs
@@ -12,9 +12,9 @@ public class Ingredient
     public int unlock;
     public string detail;
     public int degree;
-    public double bXsX;
-    public double bOsX;
-    public double bXsO;
-    public double bOsO;
+    public float bXsX;
+    public float bOsX;
+    public float bXsO;
+    public float bOsO;
 }
